Decouple unit history leg geometry from OpenLayers
Keep leg-geometry types and helpers engine-neutral in src/geo/history.ts
(OL getDistance replaced with the turf-based haversine); move OL styles,
layers and features to src/geo/engines/openlayers/unitHistoryOl.ts.
Drop unitStyles and history from PENDING_OPENLAYERS_DECOUPLE.
