+++ eslint.config.ts
@@ -23,7 +23,7 @@ const PENDING_OPENLAYERS_DECOUPLE = [
   "src/components/injects.ts",
   "src/components/commandPalette/CommandPalettePlaceItem.vue",
   "src/composables/scenarioActions.ts",
-  "src/geo/{history,history.legGeometry.test,unitStyles,utils}.ts",
+  "src/geo/utils.ts",
   "src/importexport/export/locationFormat.ts",
   "src/modules/scenarioeditor/{featureLayerUtils,referenceFeatureUtils}.ts",
   "src/modules/scenarioeditor/MapEditorMeasurementToolbar.vue",
